pyproject.toml
[project]
name            = "latex-table-generator"
description     = "Auto generator latex to table image"
readme          = "README.md"
requires-python = ">=3.10"
dependencies    = [
    "camelot-py[base]>=0.9.0",
    "img2table>=1.4.0",
    "matplotlib>=3.10.0",
    "pdf2image>=1.17.0",
    "pillow>=11.0.0",
]
dynamic         = ["version"]

[tool]
    [tool.ruff]
    target-version = "py310"
    line-length    = 130
    ignore         = ["E402", "F841", "F401", "E203"]
    [tool.hatch.version]
    path = "src/latex_table_generator/__init__.py"
